fix(internal/output): fully implement io.ReaderAt
ReadAt should block until the provided buffer is able to be fully
written to or until the underlying data structure is closed.

This also has the side effect of being able to remove output's Wait
function. Now, wait is only invoked if there are not enough bytes
available to write and the output is still open.